Fix filtering with Merged Items.
Filtering searches the current list for matching entries and keeps those. But because we had changed the path of grouped items (and possibly back again due to loading background info) it will not find a match. So instead we check through all of the items making up the merged item which have their original paths.