Flattens existing commits and updates to the Jarvis code base
* Merges duplicate movies and tv episodes into single item controllable in settings -> video -> library.

Based upon a17bdc35d8c200ac4ca725f46f068bcce3204595

Due to a couple of conflicting changes merging was not overly straight forward hence this is manually applied.
Given the age of the last commits and that much of the original patch has been changed (and will likely again) keeping this history is messy and extra work.